+++ CHANGES.md
@@ -6,6 +6,7 @@ Change Log
 #### next release (8.0.0-alpha.81)
 
 * Fix invalid HTML in `DataPreviewSections`.
+* Fix pluralisation of mapDataState to support other languages.
 * Fix CSW `Stratum` name bug.
 * Add `TerriaErrorSeverity` enum, values can be `Error` or `Warning`.
   * Errors with severity `Error` are presented to the user. `Warning` will just be printed to console.
@@ -21,6 +22,7 @@ Change Log
 * `AsyncLoader` loadXXX methods now return `Result` with `errors`.
 * Fix `MagdaReference` `forceLoadReference` bug.
 * Clean up `CkanCatalogGroup` loading - errors are no-longer swallowed.
+* Add `#configUrl` hash parameter for **dev environment only**. It can be used to overwrite Terria config URL.
 * [The next improvement]
 
 #### 8.0.0-alpha.80